Refactor C_scalar_field modules and update imports
Moved C_scalar_field medium and wave engine modules into the xperiments/C_scalar_field directory and updated all relevant imports. Renamed instrumentation module for consistency. Also standardized GUI window titles in experiment launchers by removing method suffixes.
